requirements.txt
flask
openai
requests
PyGithub
python-dotenv
